fix: consolidate to shared nft component
