+++ src/nft/components/NFTMint.tsx
@@ -1,7 +1,5 @@
-import { background, border, cn, color } from '../../styles/theme';
-import { useIsMounted } from '../../useIsMounted';
-import { useTheme } from '../../useTheme';
 import { LifecycleType, type NFTMintReact } from '../types';
+import { NFT } from './NFT';
 import { NFTLifecycleProvider } from './NFTLifecycleProvider';
 import { NFTMintProvider } from './NFTMintProvider';
 import { NFTProvider } from './NFTProvider';
@@ -18,15 +16,6 @@ export function NFTMint({
   onError,
   onSuccess,
 }: NFTMintReact) {
-  const componentTheme = useTheme();
-
-  const isMounted = useIsMounted();
-
-  // prevents SSR hydration issue
-  if (!isMounted) {
-    return null;
-  }
-
   return (
     <NFTLifecycleProvider
       type={LifecycleType.MINT}
@@ -43,20 +32,7 @@ export function NFTMint({
           useNFTMintData={useNFTMintData}
           buildMintTransaction={buildMintTransaction}
         >
-          <div
-            className={cn(
-              componentTheme,
-              color.foreground,
-              background.default,
-              border.defaultActive,
-              border.radius,
-              'flex w-full max-w-[500px] flex-col border px-6 py-4',
-              className,
-            )}
-            data-testid="ockNFTMint_Container"
-          >
-            {children}
-          </div>
+          <NFT className={className}>{children}</NFT>
         </NFTMintProvider>
       </NFTProvider>
     </NFTLifecycleProvider>
