+++ src/nft/components/NFTView.tsx
@@ -1,7 +1,5 @@
-import { background, border, cn, color } from '../../styles/theme';
-import { useIsMounted } from '../../useIsMounted';
-import { useTheme } from '../../useTheme';
 import { LifecycleType, type NFTViewReact } from '../types';
+import { NFT } from './NFT';
 import { NFTLifecycleProvider } from './NFTLifecycleProvider';
 import { NFTProvider } from './NFTProvider';
 
@@ -15,15 +13,6 @@ export function NFTView({
   onError,
   onSuccess,
 }: NFTViewReact) {
-  const componentTheme = useTheme();
-
-  const isMounted = useIsMounted();
-
-  // prevents SSR hydration issue
-  if (!isMounted) {
-    return null;
-  }
-
   return (
     <NFTLifecycleProvider
       type={LifecycleType.VIEW}
@@ -36,20 +25,7 @@ export function NFTView({
         tokenId={tokenId}
         useNFTData={useNFTData}
       >
-        <div
-          className={cn(
-            componentTheme,
-            color.foreground,
-            background.default,
-            border.defaultActive,
-            border.radius,
-            'flex w-full max-w-[500px] flex-col border px-6 pt-6 pb-4',
-            className,
-          )}
-          data-testid="ockNFTView_Container"
-        >
-          {children}
-        </div>
+        <NFT className={className}>{children}</NFT>
       </NFTProvider>
     </NFTLifecycleProvider>
   );
